# FineFundamental memory footprint improvements