--- a/Tests/Common/Data/Fundamental/MultiPeriodFieldTests.cs
+++ b/Tests/Common/Data/Fundamental/MultiPeriodFieldTests.cs
@@ -14,6 +14,7 @@
 */
 
 using System.Collections.Generic;
+using System.Linq;
 using NUnit.Framework;
 using QuantConnect.Data.Fundamental;
 
@@ -80,6 +81,119 @@ namespace QuantConnect.Tests.Common.Data.Fundamental
 
             Assert.AreEqual(1, field);
             Assert.AreEqual(1, field.Value);
+        }
+
+        [Test]
+        public void EmptyStore()
+        {
+            var field = new NormalizedBasicEPSGrowth();
+            Assert.IsFalse(field.HasValue);
+            Assert.AreEqual(0, field.Value);
+            Assert.AreEqual(0, field.FiveYears);
+            Assert.AreEqual(0, field.OneYear);
+            Assert.AreEqual(Enumerable.Empty<string>(), field.GetPeriodNames());
+            Assert.AreEqual(0,
+                field.GetPeriodValue(QuantConnect.Data.Fundamental.Period.OneYear));
+            Assert.AreEqual(0,
+                field.GetPeriodValue(QuantConnect.Data.Fundamental.Period.TenYears));
+            Assert.AreEqual(0, field.GetPeriodValues().Count);
+            Assert.IsFalse(field.HasPeriodValue(QuantConnect.Data.Fundamental.Period.OneYear));
+            Assert.IsFalse(field.HasPeriodValue(QuantConnect.Data.Fundamental.Period.TenYears));
+        }
+
+        [Test]
+        public void EmptyStoreSetPeriodValue()
+        {
+            var field = new NormalizedBasicEPSGrowth();
+            // add the default value
+            field.SetPeriodValue(QuantConnect.Data.Fundamental.Period.OneYear, 1);
+
+            Assert.IsTrue(field.HasValue);
+            Assert.AreEqual(1, field.Value);
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.OneYear, field.GetPeriodNames().Single());
+
+            var values = field.GetPeriodValues();
+            Assert.AreEqual(1, values.Count);
+            Assert.AreEqual(1, values.First().Value);
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.OneYear, values.First().Key);
+        }
+
+        [Test]
+        public void SetPeriodValue()
+        {
+            var field = new NormalizedBasicEPSGrowth();
+            // add the default value
+            field.SetPeriodValue(QuantConnect.Data.Fundamental.Period.OneYear, 1);
+            field.SetPeriodValue(QuantConnect.Data.Fundamental.Period.TenYears, 10);
+
+            Assert.IsTrue(field.HasValue);
+            Assert.AreEqual(1, field.Value);
+            var names = field.GetPeriodNames().ToList();
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.OneYear, names[0]);
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.TenYears, names[1]);
+
+            var values = field.GetPeriodValues();
+            Assert.AreEqual(2, values.Count);
+            Assert.AreEqual(1, values[QuantConnect.Data.Fundamental.Period.OneYear]);
+            Assert.AreEqual(10, values[QuantConnect.Data.Fundamental.Period.TenYears]);
+        }
+
+        [Test]
+        public void EmptyStoreUpdateValues()
+        {
+            var field = new NormalizedBasicEPSGrowth();
+
+            // update the default value
+            var data = new Dictionary<string, decimal> { { QuantConnect.Data.Fundamental.Period.OneYear, 2 } };
+            field.UpdateValues(new NormalizedBasicEPSGrowth(data));
+
+            Assert.IsTrue(field.HasValue);
+            Assert.AreEqual(2, field.Value);
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.OneYear, field.GetPeriodNames().Single());
+
+            var values = field.GetPeriodValues();
+            Assert.AreEqual(1, values.Count);
+            Assert.AreEqual(2, values.First().Value);
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.OneYear, values.First().Key);
+        }
+
+        [Test]
+        public void EmptyStoreToString()
+        {
+            var field = new NormalizedBasicEPSGrowth();
+            Assert.AreEqual("", field.ToString());
+        }
+
+        [Test]
+        public void NonEmptyStoreToString()
+        {
+            var field = new NormalizedBasicEPSGrowth();
+            field.SetPeriodValue(QuantConnect.Data.Fundamental.Period.OneYear, 1);
+            field.SetPeriodValue(QuantConnect.Data.Fundamental.Period.TenYears, 10);
+
+            Assert.AreEqual($"{QuantConnect.Data.Fundamental.Period.OneYear}:1;" +
+                            $"{QuantConnect.Data.Fundamental.Period.TenYears}:10", field.ToString());
+        }
+
+        [Test]
+        public void UpdateValues()
+        {
+            var field = new NormalizedBasicEPSGrowth();
+            // add the default value
+            field.SetPeriodValue(QuantConnect.Data.Fundamental.Period.OneYear, 1);
+
+            // update the default value
+            var data = new Dictionary<string, decimal> { { QuantConnect.Data.Fundamental.Period.OneYear, 2 } };
+            field.UpdateValues(new NormalizedBasicEPSGrowth(data));
+
+            Assert.IsTrue(field.HasValue);
+            Assert.AreEqual(2, field.Value);
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.OneYear, field.GetPeriodNames().Single());
+
+            var values = field.GetPeriodValues();
+            Assert.AreEqual(1, values.Count);
+            Assert.AreEqual(2, values.First().Value);
+            Assert.AreEqual(QuantConnect.Data.Fundamental.Period.OneYear, values.First().Key);
         }
     }
 }