… Common/Data/Fundamental/Generated/MultiPeriodValueTypes.cs