<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Will use an array as store, the most efficient collection memory wise.
Holding a custom internal `struct PeriodField` (to avoid any extra references)
- Will store period as a byte number, not a string
- By default, the collection will be null and created on demand
- Adding more unit tests

**Note 1: this PR does not change public exposed methods**
**Note 2: there is still more room for improvement, replacing stored data from decimal to double**

`Local performance tests - CoarseFineFundamentalRegressionAlgorithm 2016 to 2018 with 100 coarse symbols selected => ~35% speed improvement`
`pr`
47.77 seconds at 1k data points per second. Processing total of 51,186 data points.
48.14 seconds at 1k data points per second. Processing total of 51,186 data points.
48.21 seconds at 1k data points per second. Processing total of 51,186 data points.
`master`
73.60 seconds at 1k data points per second. Processing total of 51,186 data points.
73.95 seconds at 1k data points per second. Processing total of 51,186 data points.
76.36 seconds at 1k data points per second. Processing total of 51,186 data points.
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3471
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- Reduce memory footprint and improve speed
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit, regression and cloud testing
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->